+++ lib/features/search/sources/search_sources.dart
@@ -3,6 +3,7 @@ import 'package:flutter/widgets.dart';
 import '../models/search_source.dart';
 import 'anilist_anime_source.dart';
 import 'anilist_manga_source.dart';
+import 'comicvine_source.dart';
 import 'fantlab_source.dart';
 import 'igdb_games_source.dart';
 import 'mangabaka_source.dart';
@@ -34,6 +35,8 @@ final List<SearchSource> searchSources = List<SearchSource>.unmodifiable(
     // Books
     OpenLibrarySource(),
     FantlabSource(),
+    // Comics (a books sub-type)
+    ComicVineSource(),
   ],
 );